chore(tangle-dapp): Enable unstake page
